[create] events_locations table and function

+++ app/controllers/application_controller.rb
@@ -15,14 +15,14 @@ class ApplicationController < ActionController::Base
   end
 
   # 紐付けのないothersideを削除
-  def clean_otherside
-    user = current_user
-    user.othersides.includes(:schedules,:journals,:details).each do |otherside|
-      if otherside.schedules.empty? && otherside.journals.empty? && otherside.details.empty?
-        otherside.destroy
-      end
-    end
-  end
+  # def clean_otherside
+  #   # user = current_user
+  #   # user.othersides.includes(:schedules,:journals,:details).each do |otherside|
+  #   #   if otherside.schedules.empty? && otherside.journals.empty? && otherside.details.empty?
+  #   #     otherside.destroy
+  #   #   end
+  #   # end
+  # end
 
   # 紐付けのないmemoを削除
   def clean_memo
@@ -30,19 +30,55 @@ class ApplicationController < ActionController::Base
     memos.delete_all
   end
 
-  def crean_details #未検証要確認
+  # def crean_details #未検証要確認
+  #   # user = current_user
+  #   # user.othersides.includes(:details).each do |otherside|
+  #   #   if otherside.details.first.journal.nil?
+  #   #     otherside.details.first.destroy
+  #   #   end
+  #   #   if otherside.details.last.journal.nil?
+  #   #     otherside.details.last.destroy
+  #   #   end
+  #   # end
+  # end
+
+  def clean_parts
     user = current_user
-    user.othersides.includes(:details).each do |otherside|
-      if otherside.details.first.journal.nil?
-        otherside.details.first.destroy
-      end
-      if otherside.details.last.journal.nil?
-        otherside.details.last.destroy
+    # otherside_clear
+    user.othersides.includes(:schedules,:journals,:details).each do |otherside|
+      if otherside.schedules.empty? && otherside.journals.empty? && otherside.details.empty?
+        otherside.destroy
       end
     end
-  end
-
-  def clean_schedule_parts
+    # detail_clear 親(journals)の居ないdetailsを削除
+    user.othersides.each do |otherside|
+      if otherside.details.present?
+        if otherside.details.first.journal.nil?
+          otherside.details.first.destroy
+        end
+        if otherside.details.last.journal.nil?
+          otherside.details.last.destroy
+        end
+      end
+    end
+    # journal_clear
+    user.journals.each do |journal|
+      if journal.schedule.nil?
+        journal.destroy
+      end
+    end
+    # event_clear
+    user.events.each do |event|
+      if event.schedules.empty?
+        event.destroy
+      end
+    end
+    # location_clear
+    user.locations.each do |location|
+      if location.schedules.empty?
+        location.destroy
+      end
+    end
   end
 
   def judge_sub_or_others(otherside)
